+++ src/push/es5.js
@@ -1,5 +1,6 @@
 import { Push } from "./push";
 import { PushToken } from "./push-token";
+import { PushMessage } from "./push-message";
 
 // Declare the window object
 window.Ionic = window.Ionic || {};
@@ -7,3 +8,4 @@ window.Ionic = window.Ionic || {};
 // Ionic Namespace
 Ionic.Push = Push;
 Ionic.PushToken = PushToken;
+Ionic.PushMessage = PushMessage;