+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.6.0
+
+* **(breaking)** (push) angular integration now correctly uses notification.payload and
+  the `$stateParams` property is now passed as a standard object, not a string.
+
 ## 0.5.2
 
 * (push) Allow saving of tokens without user context